+++ src/lib/common/path.c
@@ -265,8 +265,8 @@ bp_realcost(struct as_coord from, struct as_coord to, void *pp)
 static double
 bp_seccost(struct as_coord from, struct as_coord to, void *pp)
 {
-    return (double)mapdist((coord)from.x, (coord)from.y,
-			   (coord)to.x, (coord)to.y);
+    return mapdist((coord)from.x, (coord)from.y,
+		   (coord)to.x, (coord)to.y);
 }
 
 /*
